+++ posthog-android/src/main/java/com/posthog/android/replay/PostHogSessionReplayConfig.kt
@@ -45,6 +45,22 @@ public class PostHogSessionReplayConfig
          * Defaults to 1000ms = 1s
          * Ps: it was 500ms by default until version 3.8.2
          */
-        @PostHogExperimental
+        @Deprecated("Use throttleDelayMs instead")
         public var debouncerDelayMs: Long = 1000,
-    )
+        /**
+         * Throttling delay used to reduce the number of snapshots captured and reduce performance impact
+         * This is used for capturing the view as a wireframe or screenshot
+         * The lower the number more snapshots will be captured but higher the performance impact
+         * Defaults to 1000ms = 1s
+         */
+        @PostHogExperimental
+        public var throttleDelayMs: Long = 1000,
+    ) {
+        init {
+            // for keeping back compatibility
+            @Suppress("DEPRECATION")
+            if (debouncerDelayMs != 1000L) {
+                throttleDelayMs = debouncerDelayMs
+            }
+        }
+    }